fix(server): gate launchctl tests on darwin, add timeout and cache
Move TestServerCmdIncludesLaunchctlEnv and TestMergeDarwinLaunchctlEnv
into server_darwin_test.go so the package compiles on Windows where
launchctlGetenv and mergeDarwinLaunchctlEnv are undefined.

Add a 3-second context.WithTimeout to getenvFromLaunchctl to prevent
server start from blocking indefinitely if launchctl hangs. Cache
results in a sync.Map so each key is queried at most once per process.
